refactor: MemoryMapper에 위치한 Dto 변환 로직 TimelineSliceResponse로 이동 (#358)
